+++ pydicom/config.py
@@ -223,39 +223,3 @@ def debug(debug_on=True, default_handler=True):
 
 # force level=WARNING, in case logging default is set differently (issue 103)
 debug(False, False)
-
-
-def empty_value_for_VR(VR, raw=False):
-    """Return the value for an empty element for `VR`.
-
-    The behavior of this property depends on the setting of
-    :attr:`config.use_none_as_empty_value`. If that is set to ``True``,
-    an empty value is always represented by ``None``, otherwise it depends
-    on `VR`. For text VRs (not including VRs that represent a number in
-    textual representation) an empty string is used as empty value
-    representation, for all other VRs, ``None``.
-    Note that this is used only if decoding the element - it is always
-    possible to set the value to another empty value representation,
-    which will be preserved during the element object lifetime.
-
-    Parameters
-    ----------
-    VR : str
-        The VR of the corresponding element.
-
-    raw : bool
-        If ``True``, returns the value for a :class:`RawDataElement`,
-        otherwise for a :class:`DataElement`
-
-    Returns
-    -------
-    str or bytes or None
-        The value a data element with `VR` is assigned on decoding
-        if it is empty.
-    """
-    if use_none_as_empty_text_VR_value:
-        return None
-    if VR in ('AE', 'AS', 'CS', 'DA', 'DT', 'LO', 'LT',
-              'PN', 'SH', 'ST', 'TM', 'UC', 'UI', 'UR', 'UT'):
-        return b'' if raw else ''
-    return None
